fixed width of player-card to match back button. added confirmation dialog when deleting game

+++ src/pages/game-list.tsx
@@ -22,8 +22,11 @@ export default function GameList() {
     };
 
     const deleteGame = (gameId: number) => {
-        LocalStorage.deleteGame(gameId);
-        refreshGames();
+        let answer = window.confirm("Delete game?");
+	if (answer) {
+            LocalStorage.deleteGame(gameId);
+            refreshGames();
+	}
     };
 
     const createGame = () => {
